+++ net/net.go
@@ -8,6 +8,7 @@ package gxnet
 import (
 	"net"
 	"strconv"
+	"strings"
 )
 
 // HostAddress composes a ip:port style address. Its opposite function is net.SplitHostPort.
@@ -39,10 +40,24 @@ func HostPort(addr string) (string, string, error) {
 	return net.SplitHostPort(addr)
 }
 
-func IsUDPAddrEqual(addr0 *net.UDPAddr, addr1 *net.UDPAddr) bool {
-	if addr0 != nil && addr1 != nil && addr0.IP.Equal(addr1.IP) && addr0.Port == addr1.Port && addr0.Zone == addr1.Zone {
+// refers from https://github.com/facebookgo/grace/blob/master/gracenet/net.go#L180:6
+func IsSameAddr(a1, a2 net.Addr) bool {
+	if a1.Network() != a2.Network() {
+		return false
+	}
+	a1s := a1.String()
+	a2s := a2.String()
+	if a1s == a2s {
 		return true
 	}
 
-	return false
+	// This allows for ipv6 vs ipv4 local addresses to compare as equal. This
+	// scenario is common when listening on localhost.
+	const ipv6prefix = "[::]"
+	a1s = strings.TrimPrefix(a1s, ipv6prefix)
+	a2s = strings.TrimPrefix(a2s, ipv6prefix)
+	const ipv4prefix = "0.0.0.0"
+	a1s = strings.TrimPrefix(a1s, ipv4prefix)
+	a2s = strings.TrimPrefix(a2s, ipv4prefix)
+	return a1s == a2s
 }